AF request now saving data to CoreData Friend, Photo and Group controllers now get data from CoreData

+++ GB_Vkontakte_AlexeyZhadaev/Services/Session.swift
@@ -13,7 +13,7 @@ class Session {
     static let instance = Session()
     
     var token: String?
-    var userId: Int?
+    var userId: Int64?
     
     private init() {}
 }
